fix issue: moved RestPageImpl to common-service, import RestPageImpl to OrderControllerTest and FavoritesServiceApplicationTests
