+++ commons/build.gradle
@@ -1,8 +1,68 @@
 plugins {
     id 'java'
+    id 'org.springframework.boot' version '3.0.2'
+    id 'io.spring.dependency-management' version '1.1.0'
+}
+
+ext {
+    set('mapstruct', '1.5.3.Final')
+    set('openapi', '2.0.2')
+    set('commonslang3', '3.12.0')
+    set('projectlombok', '1.18.24')
+    set('testcontainersjunit', '1.17.6')
+    set('testcontainerspostgresql', '1.17.6')
+}
+
+group = 'com.yandex.market'
+version = '0.0.1-SNAPSHOT'
+sourceCompatibility = '17'
+
+configurations {
+    compileOnly {
+        extendsFrom annotationProcessor
+    }
+}
+
+repositories {
+    mavenCentral()
 }
 
 dependencies {
-    implementation 'org.projectlombok:lombok:1.18.24'
-    implementation 'jakarta.persistence:jakarta.persistence-api:3.1.0'
+    // spring
+    implementation 'org.springframework.boot:spring-boot-starter-actuator'
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    developmentOnly 'org.springframework.boot:spring-boot-devtools'
+
+    // mapper
+    implementation "org.mapstruct:mapstruct:${mapstruct}"
+
+    // open api
+    implementation "org.springdoc:springdoc-openapi-starter-webmvc-ui:${openapi}"
+
+    // utils
+    implementation "org.apache.commons:commons-lang3:${commonslang3}"
+
+    // lombok
+    compileOnly 'org.projectlombok:lombok'
+    testCompileOnly "org.projectlombok:lombok:${projectlombok}"
+    testAnnotationProcessor "org.projectlombok:lombok:${projectlombok}"
+
+    // db
+    runtimeOnly 'org.postgresql:postgresql'
+
+    // monitoring
+    runtimeOnly 'io.micrometer:micrometer-registry-prometheus'
+
+    // annotation processor
+    annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
+    annotationProcessor 'org.projectlombok:lombok'
+    annotationProcessor "org.mapstruct:mapstruct-processor:${mapstruct}"
+
+    // tests
+    testImplementation 'org.springframework.boot:spring-boot-starter-test'
+    testImplementation 'org.springframework.amqp:spring-rabbit-test'
+    testImplementation "org.testcontainers:junit-jupiter:${testcontainersjunit}"
+    testImplementation "org.testcontainers:postgresql:${testcontainerspostgresql}"
 }
