+++ favorites-service/build.gradle
@@ -34,6 +34,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-web'
     developmentOnly 'org.springframework.boot:spring-boot-devtools'
+    testImplementation project(path: ':commons')
 
     // mapper
     implementation "org.mapstruct:mapstruct:${mapstruct}"
