+++ favorites-service/src/test/resources/db/insertFavoritesQuery.sql
@@ -12,6 +12,6 @@ INSERT INTO favorites(id,
                 user_id,
                 addition_timestamp)
 VALUES (1,'5728bd51-996c-4ddf-a97d-57855203220d',
-        '5728bd51-996c-4ddf-a97d-57855203720d',
+        '5728bd51-996c-4ddf-a97d-57355203720d',
         '5728bd51-996c-4ddf-a97d-57855203450d',
         '2023-01-13T10:14:33');
