+++ order-service/src/test/java/com/yandex/market/orderservice/OrderControllerTest.java
@@ -1,15 +1,13 @@
 package com.yandex.market.orderservice;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.yandex.market.orderservice.dto.OrderPreviewDto;
 import com.yandex.market.orderservice.dto.OrderResponseDto;
 import com.yandex.market.orderservice.model.OrderStatus;
 import com.yandex.market.orderservice.model.PaymentType;
 import com.yandex.market.orderservice.service.OrderService;
+import com.yandex.market.util.RestPageImpl;
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
@@ -17,9 +15,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.jdbc.Sql;
@@ -30,8 +25,6 @@ import org.testcontainers.junit.jupiter.Testcontainers;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.UUID;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
@@ -54,13 +47,15 @@ public class OrderControllerTest {
     @Transactional
     public void create() throws Exception {
         UUID userExternalId = UUID.fromString("cd8ae5aa-ebea-4922-b3c2-8ba8a296ef04");
-        MvcResult mvcResult = mockMvc.perform(post("/public/api/v1/users/{userId}/orders", userExternalId)
+        MvcResult mvcResult = mockMvc.perform(
+                post("/public/api/v1/users/{userId}/orders", userExternalId)
                         .contentType(MediaType.APPLICATION_JSON)
                         .content(Files.readString(Path.of("src/test/resources/CreateOrderRequestDto.json"))))
                 .andExpect(status().isCreated())
                 .andReturn();
 
-        UUID actualOrderExternalId = objectMapper.readValue(mvcResult.getResponse().getContentAsString(), UUID.class);
+        UUID actualOrderExternalId = objectMapper.readValue(mvcResult.getResponse().getContentAsString(),
+                UUID.class);
 
         Assertions.assertNotNull(orderService.getOrderResponseDtoByExternalId(actualOrderExternalId));
     }
@@ -86,7 +81,8 @@ public class OrderControllerTest {
                 .andExpect(status().isOk())
                 .andReturn();
 
-        OrderResponseDto orderResponseDto = objectMapper.readValue(mvcResult.getResponse().getContentAsString(), OrderResponseDto.class);
+        OrderResponseDto orderResponseDto = objectMapper.readValue(
+                mvcResult.getResponse().getContentAsString(), OrderResponseDto.class);
         UUID expectedUserId = UUID.fromString("cd8ae5aa-ebea-4922-b3c2-8ba8a296ef04");
         UUID actualUserId = orderResponseDto.userId();
 
@@ -113,7 +109,9 @@ public class OrderControllerTest {
                         .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andReturn();
-        Page<OrderPreviewDto> previewDto = objectMapper.readValue(mvcResult.getResponse().getContentAsString(), new TypeReference<RestPageImpl<OrderPreviewDto>>() {
+        Page<OrderPreviewDto> previewDto = objectMapper
+                .readValue(mvcResult.getResponse().getContentAsString(),
+                        new TypeReference<RestPageImpl<OrderPreviewDto>>() {
         });
 
         UUID expectedOrderExternalId = UUID.fromString("37678201-f3c8-4d5c-a628-2344eef50c54");
@@ -145,7 +143,8 @@ public class OrderControllerTest {
                 .andExpect(status().isNoContent())
                 .andReturn();
 
-        Assertions.assertEquals(OrderStatus.CANCELED, orderService.getOrderResponseDtoByExternalId(orderExternalId).orderStatus());
+        Assertions.assertEquals(OrderStatus.CANCELED,
+                orderService.getOrderResponseDtoByExternalId(orderExternalId).orderStatus());
     }
 
     @Test
@@ -200,35 +199,5 @@ public class OrderControllerTest {
                         .content(Files.readString(Path.of("src/test/resources/CreateOrderRequestDto.json"))))
                 .andExpect(status().isBadRequest())
                 .andReturn();
-    }
-
-    static class RestPageImpl<T> extends PageImpl<T> {
-
-        @JsonCreator(mode = JsonCreator.Mode.PROPERTIES)
-        public RestPageImpl(@JsonProperty("content") List<T> content,
-                            @JsonProperty("number") int number,
-                            @JsonProperty("size") int size,
-                            @JsonProperty("totalElements") Long totalElements,
-                            @JsonProperty("pageable") JsonNode pageable,
-                            @JsonProperty("last") boolean last,
-                            @JsonProperty("totalPages") int totalPages,
-                            @JsonProperty("sort") JsonNode sort,
-                            @JsonProperty("first") boolean first,
-                            @JsonProperty("numberOfElements") int numberOfElements) {
-
-            super(content, PageRequest.of(number, size), totalElements);
-        }
-
-        public RestPageImpl(List<T> content, Pageable pageable, long total) {
-            super(content, pageable, total);
-        }
-
-        public RestPageImpl(List<T> content) {
-            super(content);
-        }
-
-        public RestPageImpl() {
-            super(new ArrayList<>());
-        }
     }
 }
